pyproject.toml
[tool.poetry]
name = "dns-provisioning-job-queue"
version = "0.1.0"
description = "Control-plane style DNS provisioning with MongoDB + job queue."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "dns_queue", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
pymongo = "^4.9.0"
boto3 = "^1.35.0"  # optional, only used if you choose Route53 provider later

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
